build(fuzz): batch 2 harnesses (pic, cb8, raw, fnt, seq, audio) + the 5 bugs they found (#117) (#383)
Six libFuzzer harnesses for the image/A-V codecs, same bar as batch 1
(seed corpus, dicts, 60s ctest smoke, weekly deep run). They immediately
found five real memory-safety bugs, all fixed here with a Catch2
regression test each:

- pic_decode: TWO 32-bit-overflow bounds checks — pixels_offset + w*h and
  palette_offset + palette_size both wrapped, letting an out-of-range
  offset pass and read outside the buffer. Now computed in 64-bit.
- cb8_decode_frame: the mode bitmap is read a u32 word at a time, but the
  guard only ensured enough bits (S*8 >= cells); a 1-byte bitmap backed a
  4-byte read. Guard now requires whole words.
- fnt_repack: the glyph-body walk read cs.data[pc+1]/[pc+2] for its
  displacement disambiguation without bounds-checking, walking off a body
  truncated at the section end. Reads are bounded; a body with no RET is
  rejected.
- wav_to_pcm: a "fmt " chunk claiming size >= 16 but truncated read 16
  bytes past the file. The read is now gated on the body fitting.

fuzz_seq asserts a property, not a crash: SEQ's serializer normalizes line
endings to CRLF, so byte-identity holds only for canonical input — the
invariant for arbitrary bytes is idempotency (serialize is a fixed point
of parse). Harness and a regression test assert that.

Hardening confirmed by a 24-core fork-mode pass (6 harnesses x fork=4 x
300s, ~2 core-hours): zero findings. Full suite 313/313 with the install
censuses; specs' fuzz: lists updated; STATUS regenerated.

+++ lib/src/pic.cpp
@@ -77,7 +77,10 @@ std::vector<uint8_t> pic_decode(const uint8_t* data, size_t size,
 
     // Build effective palette: sys_pal base + inline fragment overlay
     Palette pal = sys_pal ? *sys_pal : Palette{};
-    if (info.palette_size > 0 && info.palette_offset + info.palette_size <= size) {
+    // 64-bit: palette_offset (u32) + palette_size (u32) can wrap 32 bits,
+    // passing an out-of-range offset into the read below (fuzz_pic).
+    if (info.palette_size > 0 &&
+        (uint64_t)info.palette_offset + info.palette_size <= size) {
         const uint8_t* p = data + info.palette_offset;
         int count = (int)std::min(info.palette_size / 3u, 256u);
         for (int i = 0; i < count; i++) {
@@ -104,7 +107,9 @@ std::vector<uint8_t> pic_decode(const uint8_t* data, size_t size,
     };
 
     if (fmt == 0) {
-        if (info.pixels_offset + (uint32_t)(w * h) > size) return {};
+        // 64-bit math: pixels_offset (u32) + w*h can wrap 32 bits, letting an
+        // out-of-range offset slip past the guard (fuzz_pic).
+        if ((uint64_t)info.pixels_offset + (uint64_t)w * h > size) return {};
         const uint8_t* pix = data + info.pixels_offset;
         for (int y = 0; y < h; y++)
             for (int x = 0; x < w; x++)
